Define overall project structure, and main application components
Define essential aspects of the project structure, including directory
tree, assets location (basically images, svg, videos,
and fonts files), and, main layout and most page components.

Define a 'folder per component' directory tree style. Build and style
layout components, which comprehend an inner layout component, an
outer layout component, and the header, footer, and view-more section
components as well. Build and style the home page, and other pages
components. Define alone (not styling or complete implementation) some
components for further implementation, like the image, image carousel,
video, and page templates components.

Implement routing for all pages (fully implemented or not).

Implement internationalization feature for Spanish, and English,
generating the translated text using Google Translator alone, and
therefore subject to manual review.

In the need for some components to be present, but not fully implemented
 yet, define image, image carousel, video and error pages.

Define seals and brands page templates, which are meant to be reused in
the implementation of seals and brands pages individually. Although not
implemented, the routes for each of this pages are already defined in
the the project routing.

Apply styles to each component for only the minimun and maximun viewport
sizes. Apply only layout related styles to sizes in between.